web_app/requirements.txt
googlesearch_python==1.2.3
openai==0.27.7
pandas==2.0.1
path.py==12.5.0
Requests==2.31.0
streamlit==1.22.0
